http/h2_connection: Make sure h2_connection:close() actually closes the connection when called from inside the internal cqueue
The cqueues.poll() calls in h2_connection:close() meant that :close would yield to the current controller. If the current controller was the h2_connection's own cqueue, then nothing would happen until the next action was performed on the h2_connection.
As :close() is usually the *last* thing you do to the h2_connection object, this meant that the socket wouldn't be closed until the garbage collector got around to it.

This fix is to check for a 'close_me' flag upon exit from :step() or :loop() and do the :close() then.
We still need to call cqueues.poll() twice incase other cqueues were waiting on the file descriptor